+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Correção: Corrigido tipo de Reserva "Assembleia".
 * Funcionalidade: Adicionada exibição de reservas de acordo com o auditório selecionado.
 * Funcionalidade: Adicionada função de abrir modal de reserva com horário de início preenchido.
+* Funcionalidade: Adicionada confirmação sobre termos de Local para criação de Reserva.
 
 # 0.0.2
 * Início versionamento.